KNN vs SVR for Automatic Essay Grading
In this experiment, the author makes an automated essay grading system using non-ANN algorithms, namely KNN and SVR. Before doing model building on the input data, the author performs exploratory data analysis and data preprocessing. This stage is carried out because the form of structured data can facilitate the NLP (Natural Language Processing) process, this stage is also a stage that can also affect the final result because the input data to be processed using the machine learning algorithm model is in accordance with the NLP process. After doing EDA and preprocessing, the author does model building using 2 machine learning algorithms. The KNN algorithm produces a larger MSE, so it can be said that the SVR model is more suitable for this problem. The MSE result generated by the SVR model is 0.045